fix(REACH-374): single-embed => live embed rename
