+++ packages/embed/src/constants.ts
@@ -3,7 +3,7 @@ export const POPUP_ATTRIBUTE = 'data-tf-popup'
 export const SLIDER_ATTRIBUTE = 'data-tf-slider'
 export const WIDGET_ATTRIBUTE = 'data-tf-widget'
 export const SIDETAB_ATTRIBUTE = 'data-tf-sidetab'
-export const SINGLE_EMBED_ATTRIBUTE = 'data-tf-live'
+export const LIVE_EMBED_ATTRIBUTE = 'data-tf-live'
 export const SLIDER_POSITION = 'right'
 export const SLIDER_WIDTH = 800
 export const POPUP_SIZE = 100
