+++ packages/embed/src/initializers/index.ts
@@ -3,4 +3,4 @@ export * from './initialize-popups'
 export * from './initialize-sliders'
 export * from './initialize-widgets'
 export * from './initialize-sidetabs'
-export * from './initialize-single-embeds'
+export * from './initialize-live-embeds'
