+++ packages/embed/src/live-embed/fetch-live-embed.ts
@@ -1,4 +1,4 @@
-export const fetchSingleEmbed = async (embedId: string) => {
+export const fetchLiveEmbed = async (embedId: string) => {
   const response = await fetch(`https://api.typeform.com/single-embed/${embedId}`)
 
   if (!response.ok) {
